requirements.txt
streamlit
python-dateutil
dateparser
pydantic
PyPDF2
openai
